Refractor phpstan level 5 (#130)
* Refractor phpstan level 5

* Fix phpstan

* Update composer & fix phpstan

* Update inc/uninstall.class.php


* Ignore errors individually

* Remove phpstan ignore comment

* Remove phpstan-ignore

* Remove phpstan-ignore

---------

+++ inc/uninstall.class.php
@@ -51,7 +51,7 @@ class PluginUninstallUninstall extends CommonDBTM
 
         foreach ($ma->getItems() as $itemtype => $data) {
             if (!in_array($itemtype, $UNINSTALL_TYPES)) {
-                return "";
+                return false;
             }
         }
 
@@ -67,7 +67,7 @@ class PluginUninstallUninstall extends CommonDBTM
                   Html::submit(_x('button', 'Post'), ['name' => 'massiveaction']);
                 return true;
         }
-        return "";
+        return false;
     }
 
 
@@ -92,7 +92,6 @@ class PluginUninstallUninstall extends CommonDBTM
                 Html::redirect(Plugin::getWebDir('uninstall') . '/front/action.php?device_type=' .
                   $itemtype . "&model_id=" . $_POST["model_id"]);
                 return;
-            break;
         }
         return;
     }
@@ -394,28 +393,30 @@ class PluginUninstallUninstall extends CommonDBTM
    **/
     public static function deleteOcsLink($computers_id)
     {
-        //@phpstan-ignore-next-line
-        $link = new PluginOcsinventoryngOcslink();
-        $link->dohistory = false;
-        $link->deleteByCriteria(['computers_id' => $computers_id]);
-
-        //@phpstan-ignore-next-line
-        $reg = new PluginOcsinventoryngRegistryKey();
-        $reg->deleteByCriteria(['computers_id' => $computers_id]);
+        if (class_exists('PluginOcsinventoryngOcslink')) {
+            $link = new \PluginOcsinventoryngOcslink();
+            $link->dohistory = false;
+            $link->deleteByCriteria(['computers_id' => $computers_id]);
+        }
+        if (class_exists('PluginOcsinventoryngRegistryKey')) {
+            $reg = new PluginOcsinventoryngRegistryKey();
+            $reg->deleteByCriteria(['computers_id' => $computers_id]);
+        }
     }
 
 
     public static function deleteRegistryKeys($computers_id)
     {
-        //@phpstan-ignore-next-line
-        $key = new PluginOcsinventoryngRegistryKey();
-        $key->deleteByCriteria(['computers_id' => $computers_id]);
+        if (class_exists('PluginOcsinventoryngRegistryKey')) {
+            $key = new PluginOcsinventoryngRegistryKey();
+            $key->deleteByCriteria(['computers_id' => $computers_id]);
+        }
     }
 
    /**
     * Remove a computer in the OCS database
     *
-    * @param computer_id the computer's ID in GLPI
+    * @param $computer_id the computer's ID in GLPI
     *
     * @return void
    **/
@@ -446,51 +447,57 @@ class PluginUninstallUninstall extends CommonDBTM
     {
         /** @var DBmysql $DB */
         global $DB;
-        //@phpstan-ignore-next-line
-        $DBocs = PluginOcsinventoryngOcsServer::getDBocs($ocs_server_id)->getDB();
+        if (class_exists('PluginOcsinventoryngOcsServer')) {
+            $DBocs = PluginOcsinventoryngOcsServer::getDBocs($ocs_server_id)->getDB();
 
-        //First try to remove all the network ports
-        $query = "DELETE
-                FROM `netmap`
-                WHERE `MAC` IN (SELECT `MACADDR`
-                                FROM `networks`
-                                WHERE `networks`.`HARDWARE_ID` = '" . $ocs_id . "')";
-        $DBocs->query($query);
+            //First try to remove all the network ports
+            $query = "DELETE
+                    FROM `netmap`
+                    WHERE `MAC` IN (SELECT `MACADDR`
+                                    FROM `networks`
+                                    WHERE `networks`.`HARDWARE_ID` = '" . $ocs_id . "')";
+            $DBocs->query($query);
 
-        $tables =  ["accesslog", "accountinfo", "bios", "controllers", "devices", "drives",
-            "download_history", "download_servers", "groups_cache", "inputs",
-            "memories", "modems", "monitors", "networks", "ports", "printers",
-            "registry", "slots", "softwares", "sounds", "storages", "videos"
-        ];
+            $tables =  ["accesslog", "accountinfo", "bios", "controllers", "devices", "drives",
+                "download_history", "download_servers", "groups_cache", "inputs",
+                "memories", "modems", "monitors", "networks", "ports", "printers",
+                "registry", "slots", "softwares", "sounds", "storages", "videos"
+            ];
 
-        foreach ($tables as $table) {
-            if (self::ocsTableExists($ocs_server_id, $table)) {
-                $query = "DELETE
-                      FROM `" . $table . "`
-                      WHERE `hardware_id` = '" . $ocs_id . "'";
-                $DBocs->query($query);
+            foreach ($tables as $table) {
+                if (self::ocsTableExists($ocs_server_id, $table)) {
+                    $query = "DELETE
+                        FROM `" . $table . "`
+                        WHERE `hardware_id` = '" . $ocs_id . "'";
+                    $DBocs->query($query);
+                }
             }
-        }
 
-        $query = "DELETE
-                FROM `hardware`
-                WHERE `ID` = '" . $ocs_id . "'";
-        $DBocs->query($query);
+            $query = "DELETE
+                    FROM `hardware`
+                    WHERE `ID` = '" . $ocs_id . "'";
+            $DBocs->query($query);
+        }
     }
 
 
     public static function ocsTableExists($ocs_server_id, $tablename)
     {
-        //@phpstan-ignore-next-line
-        $dbClient = PluginOcsinventoryngOcsServer::getDBocs($ocs_server_id);
+        if (class_exists('PluginOcsinventoryngOcsServer')) {
+            $dbClient = PluginOcsinventoryngOcsServer::getDBocs($ocs_server_id);
 
-        //@phpstan-ignore-next-line
-        if (!($dbClient instanceof PluginOcsinventoryngOcsDbClient)) {
-            return false;
+            if (
+                class_exists('PluginOcsinventoryngOcsDbClient')
+                && !($dbClient instanceof PluginOcsinventoryngOcsDbClient)
+            ) {
+                return false;
+            }
+
+            $DBocs = $dbClient->getDB();
+            return $DBocs->tableExists($tablename);
         }
 
-        $DBocs = $dbClient->getDB();
-        return $DBocs->tableExists($tablename);
+        return false;
     }
 
    /**
@@ -502,10 +509,11 @@ class PluginUninstallUninstall extends CommonDBTM
    */
     public static function deletePluginFieldsLink($itemtype, $items_id)
     {
-        $item = new $itemtype();
-        $item->getFromDB($items_id);
-        //@phpstan-ignore-next-line
-        PluginFieldsContainer::preItemPurge($item);
+        if (class_exists('PluginFieldsContainer')) {
+            $item = new $itemtype();
+            $item->getFromDB($items_id);
+            PluginFieldsContainer::preItemPurge($item);
+        }
     }
 
    /**
@@ -518,10 +526,13 @@ class PluginUninstallUninstall extends CommonDBTM
    **/
     public static function deleteFusionInventoryLink($itemtype, $items_id)
     {
-        if (function_exists('plugin_pre_item_purge_fusioninventory')) {
+        if (
+            class_exists('PluginFusioninventoryAgent')
+            && function_exists('plugin_pre_item_purge_fusioninventory')
+        ) {
             $item = new $itemtype();
             $item->getFromDB($items_id);
-            //@phpstan-ignore-next-line
+
             $agent = new PluginFusioninventoryAgent();
             $agents = $agent->getAgentsFromComputers([$items_id]);
 
@@ -534,10 +545,11 @@ class PluginUninstallUninstall extends CommonDBTM
                     $agent->deleteByCriteria(['id' => $current_agent['id']], true);
                 }
 
-                // remove licences
-                //@phpstan-ignore-next-line
-                $pfComputerLicenseInfo = new PluginFusioninventoryComputerLicenseInfo();
-                $pfComputerLicenseInfo->deleteByCriteria(['computers_id' => $items_id]);
+                if (class_exists('PluginFusioninventoryComputerLicenseInfo')) {
+                    // remove licences
+                    $pfComputerLicenseInfo = new PluginFusioninventoryComputerLicenseInfo();
+                    $pfComputerLicenseInfo->deleteByCriteria(['computers_id' => $items_id]);
+                }
             }
         }
     }
@@ -1045,8 +1057,11 @@ class PluginUninstallUninstall extends CommonDBTM
         /** @var array $UNINSTALL_TYPES */
         global $UNINSTALL_TYPES;
 
-        if (in_array($item->getType(), $UNINSTALL_TYPES)) {
-            self::showFormUninstallation($item->fields['id'], $item, Session::getLoginUserID());
+        if (
+            ($item instanceof CommonDBTM)
+            && in_array($item->getType(), $UNINSTALL_TYPES)
+        ) {
+            self::showFormUninstallation($item->getField('id'), $item, Session::getLoginUserID());
         }
         return true;
     }
